requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.51.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
